backend/requirements/base.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-dotenv
django-filter
djangorestframework-simplejwt
drf-yasg
django-cors-headers
gunicorn
whitenoise
dj-database-url